+++ src/domain/pageable.rs
@@ -1,10 +1,10 @@
 pub struct Pageable {
-    pub page: u64,
-    pub size: u64,
+    pub page: u32,
+    pub size: u32,
 }
 
 impl Pageable {
-    fn offset(&self) -> u64 {
+    fn offset(&self) -> u32 {
         self.page * self.size
     }
 }
